add retrofit ults for API connection
