+++ RestaurantApp/app/src/main/java/Adapter/ExpandableListAdapter.java
@@ -93,4 +93,5 @@ public class ExpandableListAdapter extends BaseExpandableListAdapter {
     public boolean isChildSelectable(int i, int i1) {
         return true;
     }
+
 }
